Rearranged class and projects to resolve dependency cycles
