+++ blended-container-registry/src/main/scala/blended/container/registry/internal/ContainerRegistryImpl.scala
@@ -18,9 +18,11 @@ package blended.container.registry.internal
 
 import akka.actor.ActorRef
 import blended.akka.{OSGIActor, OSGIActorConfig}
-import blended.container.registry.protocol._
+import blended.mgmt.base.json._
 import blended.persistence.protocol._
 import akka.actor.Props
+import blended.mgmt.base.UpdateContainerInfo
+import blended.mgmt.base.ContainerRegistryResponseOK
 
 object ContainerRegistryImpl {
   def props(cfg: OSGIActorConfig): Props = Props(new ContainerRegistryImpl(cfg))
@@ -37,7 +39,8 @@ class ContainerRegistryImpl(cfg: OSGIActorConfig) extends OSGIActor(cfg) {
       bundleActor("de.wayofquality.blended.persistence").map {
         case actor: ActorRef =>
           log.debug("Storing Container Information")
-          actor ! StoreObject(info)
+          // TODO FIXME: store container info
+          // actor ! StoreObject(info)
         case dlq if dlq == context.system.deadLetters => log.debug("Persistence manager not available")
       }
