+++ blended-mgmt-base/src/main/scala/blended/mgmt/base/UpdateContainerInfo.scala
@@ -1,0 +1,3 @@
+package blended.mgmt.base
+
+case class UpdateContainerInfo(info: ContainerInfo)
